pyproject.toml
[project]
name = "advanced-rag-lecture"
version = "0.1.0"
description = "Advanced RAG Architecture and Evaluation Tutorial"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mistralai>=1.0.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "streamlit>=1.30.0",
    "watchdog>=3.0.0",
]
